refactor(reconcile): use GetChartPath helper and tidy up exports
Replace inline path.Join(version, "charts", ...) calls with the existing
GetChartPath helper across all three reconcilers. Unexport getReleaseName
since it's only used within istiod.go. Move GetChartPath and its test to
common.go/common_test.go (renamed from types.go) since it's shared across
packages.
